% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_ranges}
\alias{merge_ranges}
\title{Merge peaks and regions GRanges}
\usage{
merge_ranges(peaks, regions)
}
\arguments{
\item{peaks}{A GRanges object}

\item{regions}{A GRanges object}
}
\value{
A DataFrame
}
\description{
Merge two GRanges objects by overlaps
}
\examples{
library(IRanges)

query <- IRanges(c(1, 4, 9), c(5, 7, 10))
subject <- IRanges(c(2, 2, 10), c(2, 3, 12))
mergeByOverlaps(query, subject)

}
