% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real_peaks}
\alias{real_peaks}
\title{AR peaks in LNCaP cell line}
\format{A \code{GRanges}}
\source{
\url{https://github.com/suwangbio/BETA/blob/master/BETA_test_data/
3656_peaks.bed}
}
\usage{
real_peaks
}
\description{
Androgen recepor peaks from ChIP-Seq experiment in the LNCaP cell line.
}
\examples{
# load data
data('real_peaks')

# locate the raw data
system.file('extdata', '3656_peaks.bed.gz', package = 'target')

# locate the source code for preparing the data
system.file('extdata', 'make-data.R', package = 'target')
}
\seealso{
\code{\link{real_transcripts}}

\code{\link{sim_peaks}}
}
\keyword{datasets}
