% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_peaks}
\alias{sim_peaks}
\title{Simulated peaks}
\format{A \code{\link[GenomicRanges]{GRanges}}}
\usage{
sim_peaks
}
\description{
is randomly generated peaks with random distances from the transcripts start
sites (TSS) of chromosome 1 of the mm10 mouse genome.
}
\examples{
\dontrun{
# load data
data('sim_peaks')

# locate the source code for preparing the data
system.file('extdata', 'make-data.R', package = 'target')
}
}
\seealso{
\code{\link{real_peaks}}

\code{\link{sim_transcripts}}
}
\keyword{datasets}
