% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_transcripts}
\alias{sim_transcripts}
\title{Simulated transcripts
The transcripts chromosome 1 of the mm10 mouse genome with randomly singed
statistics assigned to each.}
\format{A \code{GRanges}}
\usage{
sim_transcripts
}
\description{
Simulated transcripts
The transcripts chromosome 1 of the mm10 mouse genome with randomly singed
statistics assigned to each.
}
\examples{
# load data
data('sim_transcripts')

# locate the source code for preparing the data
system.file('extdata', 'make-data.R', package = 'target')
}
\seealso{
\code{\link{real_transcripts}}

\code{\link{sim_transcripts}}
}
\keyword{datasets}
